int -> float

+++ Ohjelmisto-1-PModuulit/Python - Moduuli 2 tehtävät/Tehtävä 5.py
@@ -4,4 +4,4 @@ luodit = float(input("Luodit. \n"))
 
 paino = 13.3 * (luodit + (naulat * 32 + (leiviskä * 20 * 32)))
 
-print("Massa nykymittojen mukaan: " + str(int(paino // 1000)) + "kiloa ja " + str(int(paino % 1000)) + "grammaa")
+print("Massa nykymittojen mukaan: " + str(float(paino // 1000)) + "kiloa ja " + str(float(paino % 1000)) + "grammaa")
